Move some functions into new dedicated C lib for feedback
